Move BacklogPage logic to view model
